requirements.txt
altgraph==0.17.3
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
configparser==5.3.0
docutils==0.19
grpcio==1.54.0
grpcio-tools==1.54.0
idna==3.4
importlib-metadata==6.6.0
jaraco.classes==3.2.3
joblib==1.2.0
keyring==23.13.1
markdown-it-py==2.2.0
mdurl==0.1.2
more-itertools==9.1.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pefile==2023.2.7
pkginfo==1.9.6
protobuf==3.20.3
py2exe==0.13.0.0
Pygments==2.15.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyproject_hooks==1.0.0
PythonRpcServer @ file:///C:/Users/Hp/Desktop/PythonRpcServer2
pywin32-ctypes==0.2.0
PyYAML==6.0
readme-renderer==37.3
regex==2023.3.23
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.4
six==1.16.0
tomli==2.0.1
tqdm==4.65.0
twine==4.0.2
urllib3==1.26.15
webencodings==0.5.1
zipp==3.15.0
pip~=23.1.1
wheel~=0.40.0
setuptools~=65.5.0